+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.1 (Next)
 
+* Added command line - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (1/13/2018)